requirements.txt
streamlit
pillow
pulp